refactor expression evaluation
