+++ src/vm/vm.ts
@@ -1,4 +1,13 @@
 
+// TODO:
+// VM executes a set of handlers
+// each handler is a sequence of commands, usually starting with a wait command
+// each command has a closure to keep state it needs for its execution
+// - start
+// - executeOnChange
+// - finish
+// state persisted across commands
+
 
 export function unparse(e: jsep.Expression): string {
     switch (e.type) {
@@ -35,5 +44,179 @@ export function unparse(e: jsep.Expression): string {
         }
         default:
             return "TODO"
+    }
+}
+
+
+export type GetValue = (root: string, fld: string) => any
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type StartMap = { e: jsep.Expression; v: any }[]
+
+export type CallEvaluator = (ce: jsep.CallExpression, ee: JDExprEvaluator) => any
+
+export class JDExprEvaluator {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    private exprStack: any[] = []
+
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    constructor(private env: GetValue, private callEval: CallEvaluator) {}
+
+    public tos() {
+        return this.exprStack[this.exprStack.length - 1]
+    }
+
+    public pop() {
+        return this.exprStack.pop()
+    }
+
+    public eval(e: jsep.Expression) {
+        this.exprStack = []
+        this.visitExpression(e)
+        return this.exprStack.pop()
+    }
+
+    public visitExpression(e: jsep.Expression) {
+        switch (e.type) {
+            case "ArrayExpression": {
+                // nothing to do here yet (only used for event function)
+                break
+            }
+
+            case "CallExpression": {
+                let ret = this.callEval(<jsep.CallExpression>e, this);
+                if (ret) {
+                    this.exprStack.push(ret)
+                }
+                break
+            }
+
+            case "BinaryExpression": {
+                const be = <jsep.BinaryExpression>e
+                this.visitExpression(be.left)
+                this.visitExpression(be.right)
+                const right = this.exprStack.pop()
+                const left = this.exprStack.pop()
+                switch (be.operator) {
+                    case "+":
+                        this.exprStack.push(left + right)
+                        return
+                    case "-":
+                        this.exprStack.push(left - right)
+                        return
+                    case "/":
+                        this.exprStack.push(left / right)
+                        return
+                    case "*":
+                        this.exprStack.push(left * right)
+                        return
+                    case "%":
+                        this.exprStack.push(left % right)
+                        return
+                    case ">>":
+                        this.exprStack.push(left >> right)
+                        return
+                    case ">>>":
+                        this.exprStack.push(left >>> right)
+                        return
+                    case "<<":
+                        this.exprStack.push(left << right)
+                        return
+                    case "|":
+                        this.exprStack.push(left | right)
+                        return
+                    case "&":
+                        this.exprStack.push(left & right)
+                        return
+                    case "^":
+                        this.exprStack.push(left ^ right)
+                        return
+                    case "==":
+                        this.exprStack.push(left == right)
+                        return
+                    case "!=":
+                        this.exprStack.push(left != right)
+                        return
+                    case "===":
+                        this.exprStack.push(left === right)
+                        return
+                    case "!==":
+                        this.exprStack.push(left !== right)
+                        return
+
+                    case "<":
+                        this.exprStack.push(left < right)
+                        return
+                    case ">":
+                        this.exprStack.push(left > right)
+                        return
+                    case "<=":
+                        this.exprStack.push(left <= right)
+                        return
+                    case ">=":
+                        this.exprStack.push(left >= right)
+                        return
+                }
+                break
+            }
+
+            case "UnaryExpression": {
+                const ue = <jsep.UnaryExpression>e
+                this.visitExpression(ue.argument)
+                const top = this.exprStack.pop()
+                switch (ue.operator) {
+                    case "!":
+                        this.exprStack.push(!top)
+                        return
+                    case "~":
+                        this.exprStack.push(~top)
+                        return
+                    case "-":
+                        this.exprStack.push(-top)
+                        return
+                    case "+":
+                        this.exprStack.push(+top)
+                        return
+                }
+                break
+            }
+
+            case "LogicalExpression": {
+                const le = <jsep.LogicalExpression>e
+                this.visitExpression(le.left)
+                switch (le.operator) {
+                    case "||":
+                        if (this.tos()) return
+                        else this.visitExpression(le.right)
+                        return
+                    case "&&":
+                        if (!this.tos()) return
+                        else this.visitExpression(le.right)
+                        return
+                    default:
+                }
+                break
+            }
+            case "MemberExpression": {
+                // member expressions are of form [register|event].field
+                const root = e as jsep.MemberExpression
+                const lhs = root.object as jsep.Identifier
+                const rhs = root.property as jsep.Identifier
+                const val = this.env(lhs.name, rhs.name)
+                // console.log(`${lhs.name}.${rhs.name} = ${val}`)
+                this.exprStack.push(val)
+                return
+            }
+            case "Identifier": {
+                const id = <jsep.Identifier>e
+                this.exprStack.push(this.env(id.name, ""))
+                return
+            }
+            case "Literal": {
+                const lit = <jsep.Literal>e
+                this.exprStack.push(lit.value)
+                return
+            }
+            default:
+        }
     }
 }
